requirements.txt
Django==3.2.23
djangorestframework
requests
networkx
scikit-learn
numpy